+++ testData/org/elixir_lang/code_insight/parameter_info/auto_popup_comma.ex
@@ -1,0 +1,7 @@
+defmodule ParameterInfo.AutoPopupComma do
+  def run do
+    add(1<caret>)
+  end
+
+  def add(augend, addend), do: augend + addend
+end