+++ testData/org/elixir_lang/code_insight/parameter_info/auto_popup_completion.ex
@@ -1,0 +1,5 @@
+defmodule ParameterInfo.AutoPopupCompletion do
+  def run do
+    ParameterInfo.CompletionRemote.red<caret>
+  end
+end