+++ testData/org/elixir_lang/code_insight/parameter_info/auto_popup_completion_declaration.ex
@@ -1,0 +1,4 @@
+defmodule ParameterInfo.CompletionRemote do
+  def reduce(enumerable, fun), do: {enumerable, fun}
+  def reduce_while(enumerable, acc, fun), do: {enumerable, acc, fun}
+end