+++ testData/org/elixir_lang/code_insight/parameter_info/auto_popup_opening_parenthesis.ex
@@ -1,0 +1,7 @@
+defmodule ParameterInfo.AutoPopupOpeningParenthesis do
+  def run do
+    add<caret>
+  end
+
+  def add(augend, addend), do: augend + addend
+end