+++ testData/org/elixir_lang/code_insight/parameter_info/auto_popup_remote_declaration.ex
@@ -1,0 +1,4 @@
+defmodule ParameterInfo.Remote do
+  def reduce(enumerable, fun), do: {enumerable, fun}
+  def reduce(enumerable, acc, fun), do: {enumerable, acc, fun}
+end